daemon/images: Move ErrImageDoesNotExist out of gdstore
