+++ daemon/images/errors.go
@@ -1,0 +1,23 @@
+package images
+
+import (
+	"fmt"
+
+	"github.com/distribution/reference"
+)
+
+// ErrImageDoesNotExist is error returned when no image can be found for a reference.
+type ErrImageDoesNotExist struct {
+	Ref reference.Reference
+}
+
+func (e ErrImageDoesNotExist) Error() string {
+	ref := e.Ref
+	if named, ok := ref.(reference.Named); ok {
+		ref = reference.TagNameOnly(named)
+	}
+	return fmt.Sprintf("No such image: %s", reference.FamiliarString(ref))
+}
+
+// NotFound implements the NotFound interface
+func (e ErrImageDoesNotExist) NotFound() {}
